requirements.txt
tokenizers==0.21.4
#grouped_gemm==0.1.4
peft==0.17.1
diffusers==0.33.0
decord==0.6.0
hyperpyyaml==1.2.2
soundfile==0.12.1
transformers==4.52.4
# flash_attn==2.7.0  # to accelerate installation process, please download wheel package by `wget https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.0.post1/flash_attn-2.7.0.post1+cu12torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl`
x_transformers==2.9.2
torchdiffeq
torchtune==0.6.1
torchao==0.13.0
ipynbname
accelerate==1.3.0
jiwer==3.1.0
rich
fastapi==0.118.0
uvicorn[standard]==0.37.0
